pipeline/requirements.txt
functions-framework==3.8.2
requests==2.32.3
pdfplumber==0.11.4
pydantic==2.10.0
pydantic_settings==2.6.1
google-auth==2.36.0
google-cloud-bigquery==3.27.0
google-cloud-storage==2.18.2